requirements.txt
appdirs==1.4.3
click==6.7
packaging==16.8
py==1.4.33
pyparsing==2.2.0
pytest==3.0.7
pytest-env==0.6.0
python-dotenv==0.6.3
six==1.10.0
